# Conflicts:
#	api/src/main/java/org/silsagusi/api/article/application/dto/ArticleResponse.java

+++ api/src/main/java/org/silsagusi/api/message/application/service/MessageService.java
@@ -1,15 +1,17 @@
-package org.silsagusi.api.message.application;
+package org.silsagusi.api.message.application.service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 import org.silsagusi.api.agent.infrastructure.AgentDataProvider;
 import org.silsagusi.api.customer.infrastructure.CustomerDataProvider;
 import org.silsagusi.api.message.application.dto.MessageDto;
-import org.silsagusi.api.message.application.dto.MessageUpdateRequest;
+import org.silsagusi.api.message.application.dto.UpdateMessageRequest;
+import org.silsagusi.api.message.application.mapper.MessageMapper;
 import org.silsagusi.api.message.infrastructure.dataProvider.MessageDataProvider;
+import org.silsagusi.api.message.infrastructure.validator.MessageValidator;
 import org.silsagusi.core.domain.agent.Agent;
 import org.silsagusi.core.domain.customer.entity.Customer;
+import org.silsagusi.core.domain.message.command.UpdateMessageCommand;
 import org.silsagusi.core.domain.message.entity.Message;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -26,51 +28,51 @@ public class MessageService {
 	private final AgentDataProvider agentDataProvider;
 	private final MessageDataProvider messageDataProvider;
 	private final CustomerDataProvider customerDataProvider;
+	private final MessageMapper messageMapper;
+	private final MessageValidator messageValidator;
 
 	public Page<MessageDto.Response> getMessagePage(Long agentId, Pageable pageable) {
 		Page<Message> messagePage = messageDataProvider.getMessagePageByAgent(agentId, pageable);
-		return messagePage.map(MessageDto.Response::of);
+		return messagePage.map(messageMapper::toMessageResponse);
 	}
 
-	public void createMessage(MessageDto.Request messageRequest) {
-		String content = messageRequest.getContent();
+	public void createMessages(MessageDto.Request messageRequest) {
 		List<Customer> customerList = customerDataProvider.getCustomerListByIdList(messageRequest.getCustomerIdList());
-		LocalDateTime sendAt = messageRequest.getSendAt();
 
-		customerList.forEach(customer -> {
-			Message message = new Message(
-				customer,
-				messageDataProvider.convertContent(content, customer.getName()),
-				sendAt
-			);
-			messageDataProvider.createMessage(message);
-		});
+		List<Message> messages = messageMapper.toEntityList(customerList, messageRequest.getContent(),
+			messageRequest.getSendAt());
+
+		messageDataProvider.createMessages(messages);
 	}
 
-	public void updateMessage(Long agentId, Long messageId, MessageUpdateRequest messageUpdateRequest) {
+	public void updateMessage(Long agentId, Long messageId, UpdateMessageRequest updateMessageRequest) {
 		Agent agent = agentDataProvider.getAgentById(agentId);
 		Message message = messageDataProvider.getMessage(messageId);
-		messageDataProvider.validateMessageWithAgent(message, agent);
-		messageDataProvider.validateMessageStatusEqualsPending(message);
-		message.updateMessage(messageUpdateRequest.getSendAt(), messageUpdateRequest.getContent());
-		messageDataProvider.updateMessage(message);
+
+		messageValidator.validateAgentAccess(message, agent);
+		messageValidator.validateMessageStatusEqualsPending(message);
+
+		UpdateMessageCommand updateMessageCommand = messageMapper.toUpdateMessageCommand(updateMessageRequest);
+		messageDataProvider.updateMessage(message, updateMessageCommand);
 	}
 
 	public Page<MessageDto.Response> getReservedMessagePage(Long agentId, Pageable pageable) {
 		Page<Message> messagePage = messageDataProvider.getReservedMessagePageByAgent(agentId, pageable);
-		return messagePage.map(MessageDto.Response::of);
+		return messagePage.map(messageMapper::toMessageResponse);
 	}
 
 	public MessageDto.Response getReservedMessage(Long messageId) {
 		Message message = messageDataProvider.getMessage(messageId);
-		return MessageDto.Response.of(message);
+		return messageMapper.toMessageResponse(message);
 	}
 
 	public void deleteReservedMessage(Long agentId, Long messageId) {
 		Agent agent = agentDataProvider.getAgentById(agentId);
 		Message message = messageDataProvider.getMessage(messageId);
-		messageDataProvider.validateMessageWithAgent(message, agent);
-		messageDataProvider.validateMessageStatusEqualsPending(message);
+
+		messageValidator.validateAgentAccess(message, agent);
+		messageValidator.validateMessageStatusEqualsPending(message);
+
 		messageDataProvider.deleteMessage(message);
 	}
 }
